+++ src/main/java/org/sfmis/coactivity/service/TargetGroupService.java
@@ -89,12 +89,16 @@ public class TargetGroupService extends AutoService {
 	}
 
 	public ResponseEntity<Map<String, Object>> destroy(String id) {
-		String sql = "DELETE from coactivity.targetGroup where tgId=?";
-		DbResponse rowEffect = db.execute(sql, Arrays.asList(id));
-		if (rowEffect.getErrorNumber() == 1) {
-			return Messenger.getMessenger().setMessage("Invalid Request").error();
+		if (!isBeingUsed("coactivity.targetGroup", id)) {
+			String sql = "DELETE from coactivity.targetGroup where tgId=?";
+			DbResponse rowEffect = db.execute(sql, Arrays.asList(id));
+			if (rowEffect.getErrorNumber() == 1) {
+				return Messenger.getMessenger().setMessage("Invalid Request").error();
+			} else {
+				return Messenger.getMessenger().success();
+			}
 		} else {
-			return Messenger.getMessenger().success();
+			return Messenger.getMessenger().setMessage("Deletion not allowed").error();
 		}
 	}
 }
